+++ prototypes/example_script/AP_CheckData.R
@@ -1,45 +1,45 @@
 #########################################################################################
 #
-#  Task Name:       <- Pull Water Temp check data
+#  Task Name:       <- Pull Atmospheric Pressure check data
 #
-#  Task Purpose:    <- This code pulls the Water Temp check and inspection data from both Survey123
+#  Task Purpose:    <- This code pulls the DO check and inspection data from both Survey123
 #                       and 'Provisional WaterQuality.hts' for a given site and start date
 #
-#  Task Outputs:    <- A csv of all Water Temp combined check and inspection data that is saved in
-#                       the current folder [was specified folder]
+#  Task Outputs:    <- A csv of all DO combined check and inspection data that is saved in
+#                       the specified folder
 #
-#  Notes:           <- Now reads site and dates from config.yaml in the same folder as this script,
-#                       and outputs to current folder
+#  Notes:           <-
 #
 #  Created by:      <- Hannah Marley
-#  Created on:      <- 19/01/2022
+#  Created on:      <- 02/12/2022
 #
-#  Last updated:    <- Sam Irvine
-#  Last updated by: <- 02/04/2024
+#  Last updated:    <- 05/03/2024
+#  Last updated by: <- Nic
 #
 #
 #  Written in R version 4.1.1
 #
 ########################################################################################
 
-
 # Make sure the script is working in the right place
 setwd(dirname(rstudioapi::getActiveDocumentContext()$path))
 # setwd(getSrcDirectory(function(){})[1]) # might need to use this if it's run as a script
 
-
 # load required libraries
-library(RODBC); library(dplyr); library(DT);
+library(RODBC); library(dplyr); library(DT)
 library(tidyverse); library(Hilltop); library(lubridate)
 library(yaml)
 
-rm(list = ls())
 # --------------------------------------------------------------------------------
 # --- Define site, start date, and folder file path to save inspection data to ---
 # --------------------------------------------------------------------------------
 config_yaml = yaml.load_file("./config.yaml")
 site = config_yaml$site
+
 startDate = as.Date(config_yaml$from_date)-1  # choose one day before your batch start date
+if (is.null(config_yaml$to_date)) {
+  config_yaml$to_date = format(Sys.time(), "%Y-%m-%d %H:%M:%S")
+}
 endDate = as.Date(config_yaml$to_date)+1  # choose one day after your batch end date
 folder_filepath = "./" # here
 
@@ -57,17 +57,18 @@ ch <- odbcDriverConnect('driver={SQL Server};server=DBSurvey123Live.horizons.gov
 inspections_Survey123 <- sqlQuery(ch, paste0("SELECT Hydro_Inspection.id, Hydro_Inspection.arrival_time, Hydro_Inspection.sitename,
                                              Hydro_Inspection.weather, Hydro_Inspection.notes,
                                              Hydro_Inspection.departure_time, Hydro_Inspection.creator,
-                                             DO_Inspection.inspection_id, DO_Inspection.do_notes,
-                                             WaterLevel_Inspection.inspection_id, WaterLevel_Inspection.wl_notes,
-                                             WaterTemp_Inspection.inspection_id, WaterTemp_Inspection.inspection_time, WaterTemp_Inspection.wt_device,
+                                             DO_Inspection.inspection_id, DO_Inspection.handheld_baro, DO_Inspection.logger_baro,
+                                             DO_Inspection.do_notes, DO_Inspection.inspection_id,
+                                             WaterLevel_Inspection.inspection_id,
+                                             WaterLevel_Inspection.wl_notes,
+                                             WaterTemp_Inspection.inspection_id, WaterTemp_Inspection.wt_device,
                                              WaterTemp_Inspection.handheld_temp, WaterTemp_Inspection.logger_temp
                                              FROM [dbo].Hydro_Inspection
                                              FULL JOIN [dbo].DO_Inspection ON DO_Inspection.inspection_id = Hydro_Inspection.id
                                              FULL JOIN [dbo].WaterLevel_Inspection ON WaterLevel_Inspection.inspection_id = Hydro_Inspection.id
                                              FULL JOIN [dbo].WaterTemp_Inspection ON WaterTemp_Inspection.inspection_id = Hydro_Inspection.id
                                              WHERE Hydro_Inspection.sitename = '", site, "'
-                                             AND Hydro_Inspection.arrival_time >= '", startDate, "'
-                                             AND Hydro_Inspection.arrival_time <= '", endDate, "'"),
+                                             AND Hydro_Inspection.arrival_time >= '", startDate, "'"),
                                   stringsAsFactors = F)
 
 
@@ -77,37 +78,29 @@ close(ch)
 
 # Format data
 Inspections = inspections_Survey123 %>%
-  select(id, sitename, arrival_time, inspection_time, departure_time, creator, weather,
+  select(id, sitename, arrival_time, departure_time, creator, weather,
          notes, wl_notes, wt_device, do_notes,
-         handheld_temp, logger_temp
+         handheld_baro, logger_baro
   ) %>%
-  # mutate(inspection_time = round_date(inspection_time, "15 minutes"),
-  mutate(Date = ifelse(!is.na(inspection_time),
-                       as.Date(as.character(inspection_time)),
-                               as.Date(as.character(arrival_time))),
-         # for some reason above line is displaying as numeric format, so need this next line to force to Date
-         Date = as.Date(Date),
-         Time = ifelse(!is.na(inspection_time), format(inspection_time, "%H:%M:%S"), format(arrival_time, "%H:%M:%S")))
+  mutate(arrival_time = round_date(arrival_time, "15 minutes"),
+         Date = as.Date(as.character(arrival_time)),
+         Time = format(arrival_time, "%H:%M:%S"))
 
-colnames(Inspections) = c("ID", "Site Name",
-                          "Arrival Time", "Inspection Time", "Departure Time",
-                          "InspectionStaff",
+colnames(Inspections) = c("ID", "Site Name", "Arrival Time", "Departure Time", "InspectionStaff",
                           "Weather", "Notes", "Water level notes", "MeterID", "DO Notes",
                           "Value", "Logger",
                           "Date", "Time")
 
 Inspections = Inspections %>%
   select("ID", "Site Name", "Date", "Time", "Weather",
-         "Arrival Time", "Inspection Time", "Departure Time", "InspectionStaff",
+         "Arrival Time", "Departure Time", "InspectionStaff",
          "Notes", "Water level notes", "MeterID", "DO Notes",
          "Value", "Logger") %>%
   mutate(`Arrival Time` = as.character(`Arrival Time`),
-         `Inspection Time` = as.character(`Inspection Time`),
          `Departure Time` = as.character(`Departure Time`)) %>%
   unique(.)
 
-
-write.csv(Inspections, paste0(folder_filepath, "WaterTemp_Inspections.csv"), row.names = FALSE)
+write.csv(Inspections, paste0(folder_filepath, "AP_Inspections.csv"), row.names = FALSE)
 
 
 
@@ -115,55 +108,57 @@ write.csv(Inspections, paste0(folder_filepath, "WaterTemp_Inspections.csv"), row
 
 # data source
 dat = HilltopData("//ares/Environmental Archive/Provisional WaterQuality.hts")
-meas = "Field Temperature (HRC) [Field Temperature (HRC)]"
+meas = "Field Baro Pressure (HRC) [Field Baro Pressure (HRC)]"
 #(measurements = MeasurementList(dat, site))
-
-# Get the data from Hilltop
 sites = SiteList(dat)
 
 if (site %in% sites){
+
+  # Get the data from Hilltop
   provisWQ_dat = GetData(dat, site, meas, startDate, "", WQParams = TRUE)
   provisWQ_dat = fortify.zoo(provisWQ_dat, melt = FALSE)
 
-  # --- Join all data together and format ---
-
   provisWQ_dat = provisWQ_dat %>%
-    # mutate(Index = round_date(Index, "15 minutes"),
-    mutate("Arrival Time" = as.character(Index),
-           Date = as.Date(Index),
-           Time = format(Index, "%H:%M:%S"),
-           ID = NA,
-           `Site Name` = site,
-           `Inspection Time` = NA,
-           `Departure Time` = NA,
-           `Water level notes` = NA,
-           `DO Notes` = NA,
-           `Temp Logger` = NA) %>%
-    rename("Value" = "Field Temperature (HRC)",
-           "InspectionStaff" = "SampledBy",
-           "Notes" = "Comments") %>%
-    select("ID", "Site Name", "Date", "Time", "Weather",
-           "Arrival Time", "Inspection Time", "Departure Time", "InspectionStaff",
-           "Notes", "Water level notes", "MeterID", "DO Notes",
-           "Value", "Logger") %>%
-    #rbind(., Inspections) %>%
-    #select(-ID) %>%
-    arrange(desc(`Arrival Time`))
+     mutate(Index = round_date(Index, "15 minutes"),
+            "Arrival Time" = as.character(Index),
+            Date = as.Date(Index),
+            Time = format(Index, "%H:%M:%S"),
+            ID = NA,
+            `Site Name` = site,
+            `Departure Time` = NA,
+            `Water level notes` = NA,
+            `DO Notes` = NA,
+            `Logger` = NA) %>%
+     rename("Value" = "Field Baro Pressure (HRC)",
+            "InspectionStaff" = "SampledBy",
+            "Notes" = "Comments") %>%
+     select(ID, `Site Name`, Date, Time, Weather, `Arrival Time`, `Departure Time`,
+            InspectionStaff, Notes, `Water level notes`, MeterID,
+            `DO Notes`, `Value`,
+            `Logger`) %>%
+  #   rbind(., Inspections) %>%
+  #   select(-ID) %>%
+     arrange(desc(`Arrival Time`))
   # --- save to csv ---
-  write.csv(provisWQ_dat, paste0(folder_filepath, "WaterTemp_ProvWQ.csv"), row.names = FALSE)
+  write.csv(provisWQ_dat, paste0(folder_filepath, "AP_ProvWQ.csv"), row.names = FALSE)
   full_dat = provisWQ_dat %>%
     rbind(., Inspections) %>%
     select(-ID) %>%
     arrange(desc(`Arrival Time`))
 } else {
-  full_dat = Inspections
+ full_dat = Inspections
 }
 
 
 
+
+
+
+
 # ------ Pull non-conformances for sites ------
+
 # Connect to survey123 sql table
-ch = odbcDriverConnect('driver={SQL Server};server=DBSurvey123Live.horizons.govt.nz;database=survey123;trusted_connection=true')
+ch = odbcDriverConnect('driver={SQL Server};server=sql3dev.horizons.govt.nz;database=survey123;trusted_connection=true')
 
 NCR = sqlQuery(ch, paste0("SELECT Hydro_Inspection.id, Hydro_Inspection.sitename,
                                              Non_Conformances.*
@@ -178,12 +173,11 @@ close(ch)
 # Connect to logsheets
 ch1 = odbcConnectAccess2007("//ares/HydrologySoftware/Catchment Data Tools/CDT4.accdb")
 
-NCRs = sqlQuery(ch1, paste0("SELECT NonConformance.*
-                                            FROM NonConformance
-                                            WHERE NonConformance.Sitename= '", site, "'"), stringsAsFactors = F)
-NCRs
+NCRs = sqlQuery(ch1, paste0("SELECT NonConformance.* FROM NonConformance WHERE NonConformance.Sitename= '", site, "'"),
+                stringsAsFactors = F)
 
 close(ch1)
+
 
 
 NCR_survey123 = NCR %>%
@@ -211,10 +205,9 @@ NCRs = NCRs %>%
   arrange(Date) %>%
   select(-Date) %>%
   filter(as.Date(ReportDate) >= startDate,
-         as.Date(ReportDate) <= endDate) %>%
-  unique(.)
+         as.Date(ReportDate) <= endDate)
 
-write.csv(NCRs, paste0(folder_filepath, "WaterTemp_non-conformance_reports.csv"), row.names = FALSE)
+write.csv(NCRs, paste0(folder_filepath, "AP_non-conformance_reports.csv"), row.names = FALSE)
 
 
 
@@ -222,23 +215,25 @@ write.csv(NCRs, paste0(folder_filepath, "WaterTemp_non-conformance_reports.csv")
 
 # ------ Keep just check data, format to fit Hilltop and save as csv ------
 
-# Replace "Inspections" with "full_dat" to include prov_wq in QC calculation.
-check_data = Inspections %>%
-  select(Date, Time, `Temp Check`, `Arrival Time`, `Inspection Time`, Notes) %>%
-  filter(!is.na(`Temp Check`),
+check_data = full_dat %>%
+  select(Date, Time, `Value`, `Arrival Time`, Notes) %>%
+  filter(!is.na(`Value`),
          Date <= endDate) %>%
-  mutate(`Recorder Time` = ifelse(!is.na(`Inspection Time`), `Inspection Time`, `Arrival Time`)) %>%
-  distinct(Date, `Temp Check`, .keep_all = TRUE) %>%
-  rename("Water Temperature check" = "Value",
-         "Comment" = "Notes") %>%
-  mutate(`Water Temperature check` = as.numeric(`Water Temperature check`)) %>%
-  select(Date, Time, `Water Temperature check`, `Recorder Time`, Comment) %>%
+  distinct(Date, `Value`, .keep_all = TRUE) %>%
+  dplyr::rename("Check Pressure" = "Value",
+         "Recorder Time" = "Arrival Time",
+        "Comment" = "Notes") %>%
+  mutate(`Check Pressure` = as.numeric(`Check Pressure`)) %>%
+  #select(Date, Time, `Check Pressure`, `Recorder Time`, Comment) %>%
   arrange(`Recorder Time`)
 
 # Escape newlines
 check_data$Comment = gsub("\r?\n|\r", "---NEWLINE---",check_data$Comment)
 
-write.csv(check_data, paste0(folder_filepath, "WaterTemp_check_data.csv"), row.names = FALSE)
+write.csv(check_data,
+          paste0(folder_filepath,
+                 paste0("AP_check_data_", gsub("-", "", Sys.Date()), ".csv")),
+          row.names = FALSE)